requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.161
botocore==1.29.161
channels==4.0.0
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.2
django-redis==5.3.0
django-storages==1.13.2
djangorestframework==3.14.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
redis==4.6.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
websockets==11.0.3
yarl==1.9.2